+++ pkgtmpl.py
@@ -1,22 +1,16 @@
 #!/usr/bin/env python
-"""Generate empty packages with sane defaults, for publishing to GitHub
+"""Bootstrap python packages with sane defaults
 
-Currently only supports generating python packages. Features for
-python include:
+Features include:
 
 * Generates .gitignore, AUTHORS, CHANGELOG,
   LICENSE (BSD simplified by default), MANIFEST.in, README.md,
   requirements.txt, setup.py, tests.py and tox.ini.
 * Sets your name, github username, package name in the templates by
-  reading a config that should be located @ ~/.config/pkgtmpl.ini or where
-  you specify
+  reading a config that should be located @ ~/.config/pkgtmpl.ini.
 * Optional: Add Travis CI status and twitter account to the readme
 * Sets your name and year in the copyright notice of the license file
 
-NOTE: Needs jinja2 for generating the templates. Install
-      with `pip install Jinja2`.
-TODO: Add package type specific config options, such as python versions
-      to test with in tox.
 """
 from __future__ import print_function
 import argparse
@@ -41,15 +35,19 @@ GENERAL_CONFIG_SECTION = 'general'
 FILENAMES = {
     'python': (
         '.gitignore',
+        '.travis.yml',
         'AUTHORS',
         'CHANGELOG.md',
         'LICENSE',
         'MANIFEST.in',
         'README.md',
         'requirements.txt',
+        'requirements_test.txt',
         'setup.py',
-        'tests.py',
         'tox.ini',
+        'pkgname/__init__.py',
+        'pkgname/metadata.py',
+        'pkgname/tests.py',
     )
 }
 
@@ -59,6 +57,7 @@ REQUIRED_CONFIG_VALUES = (
 )
 
 OPTIONAL_CONFIG_VALUES = (
+    'email',
     'twitter',
     'travis_ci',
 )
@@ -101,9 +100,9 @@ def validate_config_attrs(config):
                      u'{1}'.format(config._config_path, attr))
 
 
-def validate_package_path(package_path):
-    if os.path.exists(package_path):
-        sys.exit(u'Path {0} already exists.'.format(package_path))
+def validate_app_path(app_path):
+    if os.path.exists(app_path):
+        sys.exit(u'Path {0} already exists.'.format(app_path))
 
 
 def validate_package_type_exists(package_type):
@@ -115,52 +114,72 @@ def get_package_type_path(package_type):
     return os.path.join(expand_path('.'), package_type)
 
 
-def generate_files(package_type, package_path, template_context):
+def generate_files(package_type, app_path, template_context):
     package_type_path = get_package_type_path(package_type)
-    env = jinja2.Environment(loader=jinja2.FileSystemLoader(package_type_path))
-    os.mkdir(package_path)
-    print('Created directory {0}'.format(package_path))
+    env = jinja2.Environment(loader=jinja2.FileSystemLoader(package_type_path),
+                             keep_trailing_newline=True)
+    os.mkdir(app_path)
+    print('Created directory {0}'.format(app_path))
     for fn in FILENAMES[package_type]:
+
+        if fn.startswith('pkgname/'):
+            _, _, subdir_fn = fn.partition('/')
+            pkg_path = os.path.join(app_path,
+                                        template_context['package_name'])
+            print('pkg_path', pkg_path)
+            savepath = os.path.join(pkg_path, subdir_fn)
+            if not os.path.exists(pkg_path):
+                os.mkdir(pkg_path)
+        else:
+            savepath = os.path.join(app_path, fn)
+
         tmpl = env.get_template(fn)
         filecontents = tmpl.render(**template_context)
-        savepath = os.path.join(package_path, fn)
         with open(savepath, 'w') as fh:
             fh.write(filecontents)
-        print('Generated file {0}'.format(os.path.join(package_path, fn)))
+        print('Generated file {0}'.format(os.path.join(app_path, fn)))
 
 
-def git_init(package_path):
-    subprocess.call(('git', 'init'), cwd=package_path)
+def git_init(app_path):
+    subprocess.call(('git', 'init'), cwd=app_path)
 
 
-def generate(package_name, package_path, package_type,
+def generate(appname, app_path, package_type, package_name=None,
              config_path=DEFAULT_CONFIG_PATH):
-    package_path = expand_path(package_path)
+    app_path = expand_path(app_path)
     config = SimpleConfig(expand_path(config_path), GENERAL_CONFIG_SECTION)
-    config.package_name = package_name
+    config.appname = appname
+    config.package_name = package_name or appname.lower().replace('-', '_')
     validate_package_type_exists(package_type)
     validate_config_attrs(config)
-    validate_package_path(package_path)
+    validate_app_path(app_path)
     tmpl_context = {k: v for k, v in vars(config).items()
                                  if not k.startswith('_')}
     tmpl_context['today'] = datetime.now().strftime('%Y-%m-%d')
     tmpl_context['year'] = tmpl_context['today'][0:4]
 
-    generate_files(package_type, package_path, tmpl_context)
-    git_init(package_path)
+    generate_files(package_type, app_path, tmpl_context)
+    git_init(app_path)
 
 
-if __name__ == '__main__':
+def main():
     parser = argparse.ArgumentParser(
         description=__doc__,
         formatter_class=argparse.RawDescriptionHelpFormatter,
     )
-    parser.add_argument('package_name', help='The name of the package')
-    parser.add_argument('package_path', help='Path to create the project in')
-    parser.add_argument('-p', '--package-type',
+    parser.add_argument('appname', help='The name of the package')
+    parser.add_argument('app_path', help='Path to create the project in')
+    parser.add_argument('-p', '--package-name', help='Name of the package to '
+        'create inside the app dir. Defaults to the the app name in lowercase '
+        'and with dashes replaced with underscores.')
+    parser.add_argument('-P', '--package-type',
                         help='The type of package to generate',
                         default='python')
     parser.add_argument('-c', '--config-path', default=DEFAULT_CONFIG_PATH)
     args = parser.parse_args()
 
     generate(**vars(args))
+
+
+if __name__ == '__main__':
+    main()