+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.2.0
+
+**Date**: 2013-11-07
+
+* Create package directory with metadata
+* Avoid ImportErrors by using exec for getting the version info
+* Keep trailing newlines in generated files
+
 ## 0.1.0
 
 **Date**: 2013-09-14